test: :white_check_mark: landing page tests added
