+++ src/components/NoResultsFound/index.jsx
@@ -12,6 +12,9 @@ export const StyledImg = styled("img")(({ theme }) => ({
   },
 }));
 
+/**
+ * An empty state componenet telling users the system has no content to display.
+ */
 export default function NoResultsFound() {
   return (
     <Grid
