+++ src/mocks/server.js
@@ -1,0 +1,5 @@
+import { setupServer } from "msw/node";
+import { handlers } from "./handlers";
+
+// Setup requests interception using the given handlers.
+export const server = setupServer(...handlers);
